- created AttachmentsList and AttachmentsUploader and updated Attachments to use list and uploader components inside of the AttachmentsDropZone slot
